refactor(server): extract request/response types to internal/server/handlers/ package
Creates internal/server/handlers/ sub-package with 34 JSON request/response type
definitions moved out of the god package. Handler method receivers stay in
internal/server (Phase 2 work). Original handler files use type aliases to
maintain zero API-surface change.

Files created: doc.go, apikeys.go, auth.go, cache.go, itunes.go, metadata.go,
operations.go, playlists.go, reading.go

Left in place with TODO: registryProgressAdapter (has methods; used in 20+ *_ops.go files)
