+++ internal/server/handlers/doc.go
@@ -1,0 +1,8 @@
+// file: internal/server/handlers/doc.go
+// version: 1.0.0
+// guid: a1b2c3d4-e5f6-7890-abcd-ef0123456789
+// last-edited: 2026-06-01
+
+// Package handlers contains the HTTP request/response type definitions for
+// the audiobook-organizer API. Handler implementations live in internal/server.
+package handlers
