+++ internal/server/itunes_handlers.go
@@ -1,7 +1,7 @@
 // file: internal/server/itunes_handlers.go
-// version: 2.9.2
+// version: 2.10.0
 // guid: 7f2e1a4c-8b3d-4e5f-9a1b-2c3d4e5f6a7b
-// last-edited: 2026-05-19
+// last-edited: 2026-06-01
 
 // iTunes HTTP handlers. All business logic lives in internal/itunes/service.
 // Handlers that call s.itunesSvc.Importer.* guard with itunesEnabledOrError
@@ -25,6 +25,7 @@ import (
 	"github.com/jdfalk/audiobook-organizer/internal/itunes"
 	itunesservice "github.com/jdfalk/audiobook-organizer/internal/itunes/service"
 	"github.com/jdfalk/audiobook-organizer/internal/security/pathvalidation"
+	"github.com/jdfalk/audiobook-organizer/internal/server/handlers"
 	"github.com/oklog/ulid/v2"
 )
 
@@ -38,162 +39,11 @@ func (s *Server) itunesEnabledOrError(c *gin.Context) bool {
 	return true
 }
 
-// --- request / response types ---
-
-// ITunesValidateRequest represents a validation request for an iTunes library.
-type ITunesValidateRequest struct {
-	LibraryPath  string               `json:"library_path" binding:"required"`
-	PathMappings []itunes.PathMapping `json:"path_mappings,omitempty"`
-}
-
-// ITunesValidateResponse summarizes validation results for an iTunes library.
-type ITunesValidateResponse struct {
-	TotalTracks     int      `json:"total_tracks"`
-	AudiobookTracks int      `json:"audiobook_tracks"`
-	AudiobookCount  int      `json:"audiobook_count"`
-	FilesFound      int      `json:"files_found"`
-	FilesMissing    int      `json:"files_missing"`
-	MissingPaths    []string `json:"missing_paths,omitempty"`
-	PathPrefixes    []string `json:"path_prefixes,omitempty"`
-	DuplicateCount  int      `json:"duplicate_count"`
-	EstimatedTime   string   `json:"estimated_import_time"`
-}
-
-// ITunesImportRequest represents a request to import an iTunes library.
-type ITunesImportRequest struct {
-	LibraryPath      string               `json:"library_path" binding:"required"`
-	ImportMode       string               `json:"import_mode" binding:"required,oneof=organized import organize"`
-	PreserveLocation bool                 `json:"preserve_location"`
-	ImportPlaylists  bool                 `json:"import_playlists"`
-	SkipDuplicates   bool                 `json:"skip_duplicates"`
-	FetchMetadata    bool                 `json:"fetch_metadata"`
-	PathMappings     []itunes.PathMapping `json:"path_mappings,omitempty"`
-}
-
-// ITunesImportResponse acknowledges an iTunes import operation.
-type ITunesImportResponse struct {
-	OperationID string `json:"operation_id"`
-	Status      string `json:"status"`
-	Message     string `json:"message"`
-}
-
-// ITunesWriteBackRequest represents a write-back request for iTunes ITL updates.
-type ITunesWriteBackRequest struct {
-	LibraryPath  string               `json:"library_path"`
-	AudiobookIDs []string             `json:"audiobook_ids"`
-	PathMappings []itunes.PathMapping `json:"path_mappings,omitempty"`
-}
-
-// ITunesWriteBackResponse reports the result of an ITL write-back.
-type ITunesWriteBackResponse struct {
-	Success      bool   `json:"success"`
-	UpdatedCount int    `json:"updated_count"`
-	Message      string `json:"message"`
-}
-
-// ITunesBookMapping is a single book-to-iTunes-path mapping used in preview.
-//
-// Four path columns surface the full picture so users can see exactly what
-// is currently in iTunes vs what AO has on disk vs what AO would write back:
-//
-//   - ITunesPath               — what iTunes currently has, e.g. W:/foo/bar.m4b
-//   - ITunesPathTranslated     — local equivalent of ITunesPath after applying
-//     forward path mappings (so users can stat it)
-//   - AOPath                   — where AO has the file on disk (book.FilePath)
-//   - AOITunesTranslatedPath   — what AO will write into the iTunes ITL when
-//     write-back runs (ReverseRemapPath of AOPath)
-//
-// PathDiffers is true iff AOITunesTranslatedPath != ITunesPath — i.e. the
-// thing AO wants to write does not match what iTunes already has.
-//
-// Backwards compatibility: LocalPath is preserved as an alias of AOPath so
-// older clients keep working through the migration. Remove once no caller
-// reads it.
-type ITunesBookMapping struct {
-	BookID                 string `json:"book_id"`
-	Title                  string `json:"title"`
-	Author                 string `json:"author"`
-	ITunesPersistentID     string `json:"itunes_persistent_id"`
-	ITunesPath             string `json:"itunes_path,omitempty"`
-	ITunesPathTranslated   string `json:"itunes_path_translated,omitempty"`
-	AOPath                 string `json:"ao_path"`
-	AOITunesTranslatedPath string `json:"ao_itunes_translated_path,omitempty"`
-	PathDiffers            bool   `json:"path_differs,omitempty"`
-
-	// LocalPath duplicates AOPath for backwards compatibility with the
-	// previous response shape. Will be removed once no caller reads it.
-	LocalPath string `json:"local_path"`
-}
-
-// ITunesWriteBackPreviewRequest is the wire type for POST /itunes/write-back-preview.
-//
-// LibraryPath is now optional — when empty, the handler uses the configured
-// ITunesLibraryReadPath. The dialog used to require the user to type this
-// path on every preview, which was confusing because the actual write-back
-// always targets the configured ITunesLibraryWritePath (.itl) regardless.
-type ITunesWriteBackPreviewRequest struct {
-	LibraryPath string   `json:"library_path,omitempty"`
-	BookIDs     []string `json:"book_ids,omitempty"`
-}
-
-// ITunesWriteBackPreviewResponse is returned by POST /itunes/write-back-preview.
-type ITunesWriteBackPreviewResponse struct {
-	Items []ITunesBookMapping `json:"items"`
-	Total int                 `json:"total"`
-}
-
-// ITunesImportStatusResponse is returned by GET /itunes/import/:id.
-type ITunesImportStatusResponse struct {
-	OperationID string   `json:"operation_id"`
-	Status      string   `json:"status"`
-	Progress    int      `json:"progress"`
-	Message     string   `json:"message"`
-	TotalBooks  int      `json:"total_books"`
-	Processed   int      `json:"processed"`
-	Imported    int      `json:"imported"`
-	Skipped     int      `json:"skipped"`
-	Failed      int      `json:"failed"`
-	Errors      []string `json:"errors,omitempty"`
-}
-
-// ITunesTestMappingRequest tests a single path mapping against the library.
-type ITunesTestMappingRequest struct {
-	LibraryPath string `json:"library_path" binding:"required"`
-	From        string `json:"from" binding:"required"`
-	To          string `json:"to" binding:"required"`
-}
-
-// ITunesTestMappingResponse returns sample results from testing a mapping.
-type ITunesTestMappingResponse struct {
-	Tested   int                 `json:"tested"`
-	Found    int                 `json:"found"`
-	Examples []ITunesTestExample `json:"examples"`
-}
-
-// ITunesTestExample is a single found file example.
-type ITunesTestExample struct {
-	Title string `json:"title"`
-	Path  string `json:"path"`
-}
-
-// ITunesSyncRequest is the wire type for POST /itunes/sync.
-type ITunesSyncRequest struct {
-	LibraryPath  string               `json:"library_path,omitempty"`
-	PathMappings []itunes.PathMapping `json:"path_mappings,omitempty"`
-	Force        bool                 `json:"force,omitempty"`
-}
-
-// ITunesSyncResponse acknowledges a sync operation.
-type ITunesSyncResponse struct {
-	OperationID string `json:"operation_id"`
-	Message     string `json:"message"`
-}
-
 // --- handlers ---
 
 // handleITunesValidate validates an iTunes library without importing.
 func (s *Server) handleITunesValidate(c *gin.Context) {
-	var req ITunesValidateRequest
+	var req handlers.ITunesValidateRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
 		httputil.RespondWithBadRequest(c, err.Error())
 		return
@@ -217,7 +67,7 @@ func (s *Server) handleITunesValidate(c *gin.Context) {
 		return
 	}
 
-	httputil.RespondWithOK(c, ITunesValidateResponse{
+	httputil.RespondWithOK(c, handlers.ITunesValidateResponse{
 		TotalTracks:     resp.TotalTracks,
 		AudiobookTracks: resp.AudiobookTracks,
 		AudiobookCount:  resp.AudiobookCount,
@@ -232,7 +82,7 @@ func (s *Server) handleITunesValidate(c *gin.Context) {
 
 // handleITunesTestMapping tests a single path mapping against a few tracks.
 func (s *Server) handleITunesTestMapping(c *gin.Context) {
-	var req ITunesTestMappingRequest
+	var req handlers.ITunesTestMappingRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
 		httputil.RespondWithBadRequest(c, err.Error())
 		return
@@ -248,11 +98,11 @@ func (s *Server) handleITunesTestMapping(c *gin.Context) {
 		return
 	}
 
-	examples := make([]ITunesTestExample, len(resp.Examples))
+	examples := make([]handlers.ITunesTestExample, len(resp.Examples))
 	for i, e := range resp.Examples {
-		examples[i] = ITunesTestExample{Title: e.Title, Path: e.Path}
+		examples[i] = handlers.ITunesTestExample{Title: e.Title, Path: e.Path}
 	}
-	httputil.RespondWithOK(c, ITunesTestMappingResponse{
+	httputil.RespondWithOK(c, handlers.ITunesTestMappingResponse{
 		Tested:   resp.Tested,
 		Found:    resp.Found,
 		Examples: examples,
@@ -273,7 +123,7 @@ func (s *Server) handleITunesImport(c *gin.Context) {
 		return
 	}
 
-	var req ITunesImportRequest
+	var req handlers.ITunesImportRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
 		httputil.RespondWithBadRequest(c, err.Error())
 		return
@@ -316,7 +166,7 @@ func (s *Server) handleITunesImport(c *gin.Context) {
 		return
 	}
 
-	httputil.RespondWithSuccess(c, http.StatusAccepted, ITunesImportResponse{
+	httputil.RespondWithSuccess(c, http.StatusAccepted, handlers.ITunesImportResponse{
 		OperationID: op.ID,
 		Status:      "queued",
 		Message:     "iTunes import operation queued",
@@ -330,7 +180,7 @@ func (s *Server) handleITunesWriteBack(c *gin.Context) {
 		return
 	}
 
-	var req ITunesWriteBackRequest
+	var req handlers.ITunesWriteBackRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
 		httputil.RespondWithBadRequest(c, err.Error())
 		return
@@ -386,7 +236,7 @@ func (s *Server) handleITunesWriteBack(c *gin.Context) {
 	}
 
 	stdlog.Info("ITL write-back: updated tracks", "count", itlResult.UpdatedCount)
-	httputil.RespondWithOK(c, ITunesWriteBackResponse{
+	httputil.RespondWithOK(c, handlers.ITunesWriteBackResponse{
 		Success:      true,
 		UpdatedCount: itlResult.UpdatedCount,
 		Message:      fmt.Sprintf("Successfully updated %d audiobook locations in ITL", itlResult.UpdatedCount),
@@ -469,7 +319,7 @@ func (s *Server) handleITunesWriteBackPreview(c *gin.Context) {
 		return
 	}
 
-	var req ITunesWriteBackPreviewRequest
+	var req handlers.ITunesWriteBackPreviewRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
 		httputil.RespondWithBadRequest(c, err.Error())
 		return
@@ -552,7 +402,7 @@ func (s *Server) handleITunesWriteBackPreview(c *gin.Context) {
 	// historical and not worth refactoring here.
 	forwardOpts := itunes.ImportOptions{PathMappings: previewMappings}
 
-	items := make([]ITunesBookMapping, 0, len(books))
+	items := make([]handlers.ITunesBookMapping, 0, len(books))
 	for _, book := range books {
 		persistentID := *book.ITunesPersistentID
 		itunesPath := itunesLocations[persistentID]
@@ -567,7 +417,7 @@ func (s *Server) handleITunesWriteBackPreview(c *gin.Context) {
 		if itunesPath != "" {
 			itunesTranslated = forwardOpts.RemapPath(itunesPath)
 		}
-		items = append(items, ITunesBookMapping{
+		items = append(items, handlers.ITunesBookMapping{
 			BookID:                 book.ID,
 			Title:                  book.Title,
 			Author:                 author,
@@ -581,7 +431,7 @@ func (s *Server) handleITunesWriteBackPreview(c *gin.Context) {
 		})
 	}
 
-	httputil.RespondWithOK(c, ITunesWriteBackPreviewResponse{
+	httputil.RespondWithOK(c, handlers.ITunesWriteBackPreviewResponse{
 		Items: items,
 		Total: len(items),
 	})
@@ -635,7 +485,7 @@ func (s *Server) handleListITunesBooks(c *gin.Context) {
 		filtered = filtered[offset:end]
 	}
 
-	items := make([]ITunesBookMapping, 0, len(filtered))
+	items := make([]handlers.ITunesBookMapping, 0, len(filtered))
 	for _, book := range filtered {
 		author := ""
 		if book.AuthorID != nil {
@@ -643,7 +493,7 @@ func (s *Server) handleListITunesBooks(c *gin.Context) {
 				author = a.Name
 			}
 		}
-		items = append(items, ITunesBookMapping{
+		items = append(items, handlers.ITunesBookMapping{
 			BookID:             book.ID,
 			Title:              book.Title,
 			Author:             author,
@@ -678,7 +528,7 @@ func (s *Server) handleITunesImportStatus(c *gin.Context) {
 	progress := calculatePercent(op.Progress, op.Total)
 	snapshot := s.itunesSvc.Importer.GetStatus(op.ID)
 
-	httputil.RespondWithOK(c, ITunesImportStatusResponse{
+	httputil.RespondWithOK(c, handlers.ITunesImportStatusResponse{
 		OperationID: op.ID,
 		Status:      op.Status,
 		Progress:    progress,
@@ -711,7 +561,7 @@ func (s *Server) handleITunesImportStatusBulk(c *gin.Context) {
 
 	snapshots := s.itunesSvc.Importer.GetStatusBulk(req.IDs)
 
-	results := make(map[string]ITunesImportStatusResponse, len(req.IDs))
+	results := make(map[string]handlers.ITunesImportStatusResponse, len(req.IDs))
 	for _, opID := range req.IDs {
 		op, err := s.Store().GetOperationByID(opID)
 		if err != nil || op == nil {
@@ -722,7 +572,7 @@ func (s *Server) handleITunesImportStatusBulk(c *gin.Context) {
 		if snapshot == nil {
 			snapshot = &itunesservice.ImportStatusSnapshot{}
 		}
-		results[opID] = ITunesImportStatusResponse{
+		results[opID] = handlers.ITunesImportStatusResponse{
 			OperationID: op.ID,
 			Status:      op.Status,
 			Progress:    progress,
@@ -804,9 +654,9 @@ func (s *Server) handleITunesSync(c *gin.Context) {
 		return
 	}
 
-	var req ITunesSyncRequest
+	var req handlers.ITunesSyncRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
-		req = ITunesSyncRequest{}
+		req = handlers.ITunesSyncRequest{}
 	}
 
 	libraryPath := req.LibraryPath
@@ -864,7 +714,7 @@ func (s *Server) handleITunesSync(c *gin.Context) {
 		return
 	}
 
-	httputil.RespondWithSuccess(c, http.StatusAccepted, ITunesSyncResponse{
+	httputil.RespondWithSuccess(c, http.StatusAccepted, handlers.ITunesSyncResponse{
 		OperationID: op.ID,
 		Message:     "iTunes sync operation queued",
 	})
